requirements.txt
Flask==2.2.3
gunicorn==20.1.0
python-telegram-bot==13.15
requests==2.28.2